requirements.txt
einops>=0.7.0
numpy>=1.24.4
torch>=2.1.0
torchvision>=0.16.0

PyYAML>=6.0.1
